Add a 'sample' property to violation reports.
Firefox has shipped a 'script-sample' property in CSP violation reports
since their initial implementation, while other browsers have resisted
doing the same. This patch attempts to reach concensus on an opt-in
variant Firefox's behavior. In short, we'll collect a 40-character
sample for inline script and style violations, and include it in the
violation report (and associated SecurityPolicyViolationEvent object)
iff a `'report-sample'` expression is present in the violated directive.
That is, given:

    script-src 'nonce-abc' 'report-sample'

and:

    <script>alert(1);</script>

The violation report would include:

    {
      ...
      "sample": "alert(1);"
    }

Tests for this patch are underway in
https://codereview.chromium.org/2436003002.