Refactored fetch directives for readability and logic. (#318)
Updated #effective-directive-for-a-request with new directives and updates.
Updated fetch directives to use #effective-directive-for-[a-request/inline-check] always, instead of spelling out the exact same logic.

Added #effective-directive-for-inline-check which returns the effective directive for inline checks

Added a #directive-fallback-list algorithm which will return an orderered set of a directives fallback list. This keeps the fallback logic clear and centralized instead of it being entirely distributed between directives. This means there is an easy way to, for instance, figure out what directive is used as a fallback if worker-src is not present.

Added a #should-directive-execute algorithm which will take over all checks that were before distributed between all directives. It uses the newly added #directive-fallback-list algorithm.
This means that the default-src directive won't have to have tons of if checks to check if any more relevant directive is present for the particular request or inline check.
Updated all fetch directives to use #should-directive-execute.

Fixed issue where default-src would not be used as fallback for inline script/style checks according to the spec.